pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genimg"
version = "0.10.0"
description = "AI image generation tool with prompt optimization"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "codeprimate"}
]
maintainers = [
    {name = "codeprimate"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "gradio>=6.0.0,<7",
    "requests>=2.28.0",
    "pillow>=10.0.0",
    "pillow-heif>=0.13.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    # Image analysis (describe): Florence-2, JoyTag
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "huggingface_hub>=0.20.0",
    "safetensors>=0.4.0",
    "timm>=0.9.0",
    "torchvision>=0.15.0",
    "einops>=0.7.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "pytest-asyncio>=0.21.0",
    "responses>=0.24.0",
    # Code Quality
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    # Type stubs
    "types-requests>=2.31.0",
    "types-Pillow>=10.0.0",
    "types-PyYAML>=6.0.0",
    # Pre-commit
    "pre-commit>=3.5.0",
    # Build and publish
    "build>=1.2.0",
    "twine>=6.0.0",
]

[project.scripts]
genimg = "genimg.__main__:main"
genimg-ui = "genimg.ui.gradio_app:main"

[project.urls]
Homepage = "https://github.com/codeprimate/genimg"
Documentation = "https://github.com/codeprimate/genimg#readme"
Repository = "https://github.com/codeprimate/genimg"
"Bug Tracker" = "https://github.com/codeprimate/genimg/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
genimg = ["prompts.yaml", "ui_models.yaml", "assets/logo/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--cov=src/genimg",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests (real backend/API)",
    "slow: Slow tests (run with pytest --run-slow to include)",
    "expensive: Expensive tests (API cost)",
]

[tool.coverage.run]
source = ["src/genimg"]
omit = ["tests/*", "src/genimg/__main__.py", "src/genimg/core/image_analysis/backends/_joytag_models.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312", "py313"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | _joytag_models\.py
)/
'''

[tool.ruff]
target-version = "py310"
line-length = 100
exclude = [
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "src/genimg/core/image_analysis/backends/_joytag_models.py",  # vendored; do not lint/format
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "PTH", # flake8-use-pathlib
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "N806", "SIM117"]  # assert in tests; mock names (Popen); nested with
"src/genimg/ui/gradio_app.py" = ["N806", "SIM117"]  # UI constants (_UI_CONCURRENCY_ID, _JS_SET_PAGE_TITLE); nested with for Gradio layout
"src/genimg/core/image_analysis/backends/joytag.py" = ["N812"]  # TVF common alias for torchvision.transforms.functional

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
exclude = ["src/genimg/core/image_analysis/backends/_joytag_models.py"]

[[tool.mypy.overrides]]
module = [
    "gradio.*",
    "PIL.*",
    "pillow_heif.*",
    "torch.*",
    "torchvision.*",
]
ignore_missing_imports = true
